Changed data from unsigned int to long to long long

+++ src/include/interpreter/Program.hpp
@@ -9,7 +9,7 @@ namespace bs {
 
 	struct Token {
 		char identifier;
-		long data;
+		long long data;
 	};
 
 	struct Program {
